New functions for persons
Only small changes were made, allowing for more checking if any person records that are already in Iconobase have the right type of person. A new function was introduced to save Person records (and later also other records). This has actually the same purpose as the 'Create record' button, but I do not want to interfere with the latter.